Avatar + UserDisplay
Created an avatar component with a fallback that shows the first 2 letters of the user name (maybe try to change to make it first and last name letters). The user display is also now done from scratch without material UI components.

Tag is a css style for div, maybe it should be a component if we want to reuse it.

Next up is connecting the data to props like we did with events.