Props added to userDisplay
- props instead of hard coded
- asset for profile pic for avatar
- mock data for a "userA" for testing in app file

user display object also has unused parameters: country and handle. The handle should be used when we dont use the full user display. And we need to find a library where we can search for flags and add the right one (alternatively just do it the same way as profile pic)