Sidebar, Event page structure, Database connection, Event Card structure
Finished the sidebar.
Started structuring the event page.
Made a database in back4app and connected it to our app. (use import Parse from 'parse', in files you want to connect to the database)
Restructured the event card to get data from the database.
Added font for logo
Added logo as browser icon.