Profile info connected to data base
Now the side bar will show the info for the logged in user (changes when you change user)

Connected the profile display part of the side bar to the logged in user info. The data is from UserPublic, but the query uses the private _User data to identify the logged in user.

I also updated profile display to use our custom Avatar component instead of MUI. And added text-overflow behavior to the css.  (NOTE: lets consider if we can merge some of the userdisplay and profile info code since it is very similar, we should at least use the same css since it is similar - maybe it could be a variant of the same component)

I also removed the instance from the places file.

Might be relevant to move some of the parse functions to another file to make it more structured e.g. we could have a function to get the user info or create an object for a user based on some selection.