Testing out UI for filters
props are not correctly se up, hard coded labels for filters.

created a button component which I think works fine. Replaced the log out button with the new button and used the destructive variant.

We need to decide if we want to have filters like a filter bar where you can select many filters (each is a toggle) which is what I started doing now. Or if we just want to have one selected at a time more like a toggle button group so selecting one makes all the others de-selected.

Not ready to merge yet.