Filters work
Not ready, but the base functionality now works.

Do we want it to act as a toggle button so only one is selected or can you select multiple and will it then show both. Might depend on the use case (do we want it for categories like in my example here, or do we want it to apply multiple e.g. category, tag, time, and then it needs to match all the filters?)

Again, I have made some very unpolished changes to the card styling to test it. I will fix all that before trying to merge.