UI for creating a post
created toggle button group, almost the same as filters, but only one can be selected at a time (some updates to the button component as well)

Created a component: CreatePost, which opens up a dialog (maybe dialog should be its own component with a slot, if we need to be able to reuse it).

Some changes to the textfield as well but I am still struggling a bit with the layout of it.

Changed a few things from the design both because it made more sense in code but also UX. (like having a block button in the dialog and not in the container)

Not started on writing to the database but that is next step.

also made some small changes to post template to not require event icon (it became a problem only after connecting to new database but not sure why)