CreatePost (Thread) in database
Now it works to create a new post on the database. Currently only supporting threads so I have disabled the post button if the toggle button group is not selecting threads.

It also works to upload a photo to the database from the text field photo icon button. Currently I am just using a sample image and converting it into a file when uploading to parse but ) think this this will be more simple after I add the upload file functionality.

It gets the user public info about the logged in user. Although it might be better to make the app just get the user data once logged in and store it instead of calling it each time you post (then have some sort of check if the user is changed or logged out).

I also changed a bit in the home page, removed the text field and just using a button for now to open the dialog. (used a dev-description to describe some things on the screen).