Services and event input
Created a services folder and 2 files for postService and userService to make code more clean.

userService replaces long code in sideBarNav to fetch user data and is also used in postService.

Now it is also possible to create a new event, but the only input for now is the postTitle. The rest is hard coded fallbacks.

I added an Icon for the "other" category to use as a fallback, HOWEVER note that if you add a new event in the "other" category the site will crash for those who have not updated their version with the new icon....