FileUpload
Now you can upload files from your device and it will be displayed in the create post component first and then uploaded to parse with the post.

only works on events now (not with the icons...)

Also requires the image to have a .png or .jpg name.

Next up is using forward ref and  imperitive handle to make parent component be able to trigger and control the file upload.

Then I might add a function to verify the name/type/size of the file (maybe a cloud function?)