file upload and more create post work
Added a cloud function that validates if the event time is in the past and throws an error message to the UI (would like to do this for more fields to avoid issues, and maybe to some instant validation on the UI front to avoid parse errors all together)

Added a function handleClose to reset all the variables when closing the dialog.

made the dialog bigger.

Then I went a bit off track and decided to add a profile picture upload (since I had done file upload for posts). So now you can add and change a profile photo from the profile page. I also have an instance of the userDisplay (not sure if that is relevant).

Deleted the extra prop "userInfo" from userDisplay since we are no longer using hard coded data. Also deleted the data file. Also made the dot and time div conditional on the time prop so we dont show a dot with no time if no time is there.

Also created css selector for error and success message on page level. Maybe we should look at how we want to do this overall.

Missing/Todo:
-Add a remove photo from post after you have selected the photo.
-Add a rename file flow since you need to have .png, .jpg or such.
-For create post it would be nice to have a confirmation dialog before closing so you dont remove all the input accidentally by clicking outside of the dialog (maybe a draft function?)
-look into the sidebar profile display which does not update when you change the profile picture unless you refresh the screen (is that ok or do we want to fix that?)